Fix missing index content on JavaScript disabled
