--- vitepress/.vitepress/theme/NotesIndexClient.vue
@@ -1,49 +1,0 @@
-<script setup>
-import { data as importedData } from '../../notes.data'
-import { excerptToLink } from './utils/frontmatter'
-import { datetime } from 'datetime-attribute'
-
-// The following declaration is just to import type autocompletion. 😑
-
-/** @type {Record<string, import('vitepress').ContentData[]>} */
-const data = importedData
-
-// Sort years
-const years = Object.keys(data).sort((a, b) => b - a)
-
-years.forEach(year => data[year].forEach(note => {
-
-  // normalize date
-  note.frontmatter.publishedAt = new Date(note.frontmatter.publishedAt)
-
-  /**
-   * Ideally this transformation should be done at build time, but I don’t
-   * want to add a DOM Parser library because Node doesn’t have one. 😑
-   * It turns an excerpt into a proper link for the content index.
-   */
-  excerptToLink(note)
-}))
-
-const dateFormatter = new Intl.DateTimeFormat('en-GB', {
-  day: 'numeric',
-  month: 'long',
-})
-</script>
-
-<template>
-  <template v-for="year in years">
-    <h2><time :datetime="datetime(new Date(year), 'year')">{{ year }}</time></h2>
-
-    <ul>
-      <li v-for="({ frontmatter: { title, publishedAt, excerpt }, url }) in data[year]">
-        <p>
-          <time :datetime="datetime(publishedAt)">
-            {{ dateFormatter.format(publishedAt) }}
-          </time>:
-          <span v-if="excerpt" v-html="excerpt"/>
-          <a v-else :href="url">{{ title }}</a>
-        </p>
-      </li>
-    </ul>
-  </template>
-</template>
