+++ vitepress/.vitepress/theme/index.js
@@ -1,10 +1,6 @@
 import './custom.css'
 import DefaultTheme from 'vitepress/theme'
-// import NotesIndex from './NotesIndex.vue'
 
 export default {
   extends: DefaultTheme,
-  // enhanceApp({ app }) {
-  //   app.component('notes-index', NotesIndex)
-  // }
 }
